# perf: add lua iteration to decrease loops in lua