--- a/lua/ivy/libivy_test.lua
+++ b/lua/ivy/libivy_test.lua
@@ -10,18 +10,37 @@ end)
 
 it("should find a dot file", function(t)
   local current_dir = libivy.ivy_cwd()
-  local matches = libivy.ivy_files(".github/workflows/ci.yml", current_dir)
+  local results = libivy.ivy_files(".github/workflows/ci.yml", current_dir)
 
-  local results = {}
-  for line in string.gmatch(matches, "[^\r\n]+") do
-    table.insert(results, line)
+  if results.length ~= 2 then
+    t.error("Incorrect number of results found " .. results.length)
   end
 
-  if #results ~= 2 then
-    t.error "Incorrect number of results"
+  if results[2].content ~= ".github/workflows/ci.yml" then
+    t.error("Invalid matches: " .. results[2].content)
+  end
+end)
+
+it("will allow you to access the length via the metatable", function(t)
+  local current_dir = libivy.ivy_cwd()
+  local results = libivy.ivy_files(".github/workflows/ci.yml", current_dir)
+
+  local mt = getmetatable(results)
+
+  if results.length ~= mt.__len(results) then
+    t.error "The `length` property does not match the __len metamethod"
+  end
+end)
+
+it("will create an iterator", function(t)
+  local iter = libivy.ivy_files(".github/workflows/ci.yml", libivy.ivy_cwd())
+  local mt = getmetatable(iter)
+
+  if type(mt["__index"]) ~= "function" then
+    t.error "The iterator does not have an __index metamethod"
   end
 
-  if results[2] ~= ".github/workflows/ci.yml" then
-    t.error("Invalid matches: " .. results[2])
+  if type(mt["__len"]) ~= "function" then
+    t.error "The iterator does not have an __len metamethod"
   end
 end)