--- a/lua/ivy/window_test.lua
+++ b/lua/ivy/window_test.lua
@@ -21,3 +21,13 @@ it("can set items", function(t)
   window.set_items { { content = "Line one" } }
   t.assert_equal("Line one", window.get_current_selection())
 end)
+
+it("will set the items when a string is passed in", function(t)
+  window.initialize()
+
+  local items = table.concat({ "One", "Two", "Three" }, "\n")
+  window.set_items(items)
+
+  local lines = table.concat(vim_mock.get_lines()[window.buffer], "\n")
+  t.assert_equal(items, lines)
+end)